requirements.txt
pandas
scikit-learn
matplotlib
seaborn
nltk
imapclient
python-dotenv
joblib
